pyproject.toml
[project]
name = "employee-help"
version = "0.1.0"
description = "AI-powered legal guidance assistant for California employment discrimination law"
readme = "docs/requirements/PHASE_1_KNOWLEDGE_ACQUISITION.md"
requires-python = ">=3.12"
dependencies = [
    "playwright>=1.49",
    "beautifulsoup4>=4.12",
    "pdfplumber>=0.11",
    "structlog>=24.0",
    "pyyaml>=6.0",
    "lxml>=5.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=6.0",
    "pytest-asyncio>=0.24",
]

[project.scripts]
employee-help = "employee_help.cli:main"

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
addopts = [
    "--strict-markers",
    "--tb=short",
    "-q",
]
markers = [
    "integration: marks tests that hit live external services (deselect with '-m \"not integration\"')",
]

[tool.coverage.run]
source = ["src/employee_help"]
omit = ["*/cli.py"]

[tool.coverage.report]
fail_under = 80
show_missing = true

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/employee_help"]
